remove free id and set consistency with uid

+++ flappy_goud/GoudGame.cs
@@ -81,34 +81,32 @@ public class GoudGame
         }
     }
 
-    public int AddSprite(string texturePath, SpriteDto data)
-
+    public uint AddSprite(string texturePath, SpriteDto data)
     {
         unsafe
         {
             fixed (byte* texturePathBytes = System.Text.Encoding.ASCII.GetBytes(texturePath + "\0"))
             {
-                return (int)NativeMethods.game_add_sprite(gameInstance, texturePathBytes, data);
+                return NativeMethods.game_add_sprite(gameInstance, texturePathBytes, data);
             }
         }
     }
 
-    public void UpdateSprite(int id, SpriteDto data)
+    public void UpdateSprite(uint id, SpriteDto data)
     {
         unsafe
         {
-            NativeMethods.game_update_sprite(gameInstance, (uint)id, data);
+            NativeMethods.game_update_sprite(gameInstance, id, data);
         }
     }
 
-    public void RemoveSprite(int id)
+    public void RemoveSprite(uint id)
     {
         unsafe
         {
-            NativeMethods.game_remove_sprite(gameInstance, (uint)id);
+            NativeMethods.game_remove_sprite(gameInstance, id);
         }
     }
-
     public void Close()
     {
         unsafe
@@ -117,7 +115,7 @@ public class GoudGame
         }
     }
 
-    
+
